# golang/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
      "description": "This is a server for managing events. Imersão Full Cycle",
      "title": "Events API",
      "contact": {},
      "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
      "/checkout": {
          "post": {
              "description": "Buy tickets for a specific event",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "Buy tickets for an event",
              "parameters": [
                  {
                      "description": "Input data",
                      "name": "input",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/usecase.BuyTicketsInputDTO"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/usecase.BuyTicketsOutputDTO"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "type": "string"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "type": "string"
                      }
                  }
              }
          }
      },
      "/events": {
          "get": {
              "description": "Get all events with their details",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "List all events",
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/usecase.ListEventsOutputDTO"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "type": "string"
                      }
                  }
              }
          },
          "post": {
              "description": "Create a new event with the given details",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "Create a new event",
              "parameters": [
                  {
                      "description": "Input data",
                      "name": "input",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/usecase.CreateEventInputDTO"
                      }
                  }
              ],
              "responses": {
                  "201": {
                      "description": "Created",
                      "schema": {
                          "$ref": "#/definitions/usecase.CreateEventOutputDTO"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "type": "string"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "type": "string"
                      }
                  }
              }
          }
      },
      "/events/{eventID}": {
          "get": {
              "description": "Get details of an event by ID",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "Get event details",
              "parameters": [
                  {
                      "type": "string",
                      "description": "Event ID",
                      "name": "eventID",
                      "in": "path",
                      "required": true
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/usecase.GetEventOutputDTO"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "type": "string"
                      }
                  },
                  "404": {
                      "description": "Not Found",
                      "schema": {
                          "type": "string"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "type": "string"
                      }
                  }
              }
          }
      },
      "/events/{eventID}/spots": {
          "get": {
              "description": "List all spots for a specific event",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "List spots for an event",
              "parameters": [
                  {
                      "type": "string",
                      "description": "Event ID",
                      "name": "eventID",
                      "in": "path",
                      "required": true
                  }
              ],
              "responses": {
                  "200": {
                      "description": "OK",
                      "schema": {
                          "$ref": "#/definitions/usecase.ListSpotsOutputDTO"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "type": "string"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "type": "string"
                      }
                  }
              }
          },
          "post": {
              "description": "Create a specified number of spots for an event",
              "consumes": [
                  "application/json"
              ],
              "produces": [
                  "application/json"
              ],
              "tags": [
                  "Events"
              ],
              "summary": "Create spots for an event",
              "parameters": [
                  {
                      "type": "string",
                      "description": "Event ID",
                      "name": "eventID",
                      "in": "path",
                      "required": true
                  },
                  {
                      "description": "Input data",
                      "name": "input",
                      "in": "body",
                      "required": true,
                      "schema": {
                          "$ref": "#/definitions/http.CreateSpotsRequest"
                      }
                  }
              ],
              "responses": {
                  "201": {
                      "description": "Created",
                      "schema": {
                          "$ref": "#/definitions/usecase.CreateSpotsOutputDTO"
                      }
                  },
                  "400": {
                      "description": "Bad Request",
                      "schema": {
                          "$ref": "#/definitions/http.ErrorResponse"
                      }
                  },
                  "500": {
                      "description": "Internal Server Error",
                      "schema": {
                          "$ref": "#/definitions/http.ErrorResponse"
                      }
                  }
              }
          }
      }
  },
  "definitions": {
      "http.CreateSpotsRequest": {
          "type": "object",
          "properties": {
              "number_of_spots": {
                  "type": "integer"
              }
          }
      },
      "http.ErrorResponse": {
          "type": "object",
          "properties": {
              "message": {
                  "type": "string"
              }
          }
      },
      "usecase.BuyTicketsInputDTO": {
          "type": "object",
          "properties": {
              "card_hash": {
                  "type": "string"
              },
              "email": {
                  "type": "string"
              },
              "event_id": {
                  "type": "string"
              },
              "spots": {
                  "type": "array",
                  "items": {
                      "type": "string"
                  }
              },
              "ticket_type": {
                  "type": "string"
              }
          }
      },
      "usecase.BuyTicketsOutputDTO": {
          "type": "object",
          "properties": {
              "tickets": {
                  "type": "array",
                  "items": {
                      "$ref": "#/definitions/usecase.TicketDTO"
                  }
              }
          }
      },
      "usecase.CreateEventInputDTO": {
          "type": "object",
          "properties": {
              "capacity": {
                  "type": "integer"
              },
              "date": {
                  "type": "string"
              },
              "image_url": {
                  "type": "string"
              },
              "location": {
                  "type": "string"
              },
              "name": {
                  "type": "string"
              },
              "organization": {
                  "type": "string"
              },
              "partner_id": {
                  "type": "integer"
              },
              "price": {
                  "type": "number"
              },
              "rating": {
                  "type": "string"
              }
          }
      },
      "usecase.CreateEventOutputDTO": {
          "type": "object",
          "properties": {
              "capacity": {
                  "type": "integer"
              },
              "date": {
                  "type": "string"
              },
              "id": {
                  "type": "string"
              },
              "image_url": {
                  "type": "string"
              },
              "location": {
                  "type": "string"
              },
              "name": {
                  "type": "string"
              },
              "organization": {
                  "type": "string"
              },
              "partner_id": {
                  "type": "integer"
              },
              "price": {
                  "type": "number"
              },
              "rating": {
                  "type": "string"
              }
          }
      },
      "usecase.CreateSpotsOutputDTO": {
          "type": "object",
          "properties": {
              "spots": {
                  "type": "array",
                  "items": {
                      "$ref": "#/definitions/usecase.SpotDTO"
                  }
              }
          }
      },
      "usecase.EventDTO": {
          "type": "object",
          "properties": {
              "capacity": {
                  "type": "integer"
              },
              "date": {
                  "type": "string"
              },
              "id": {
                  "type": "string"
              },
              "image_url": {
                  "type": "string"
              },
              "location": {
                  "type": "string"
              },
              "name": {
                  "type": "string"
              },
              "organization": {
                  "type": "string"
              },
              "partner_id": {
                  "type": "integer"
              },
              "price": {
                  "type": "number"
              },
              "rating": {
                  "type": "string"
              }
          }
      },
      "usecase.GetEventOutputDTO": {
          "type": "object",
          "properties": {
              "capacity": {
                  "type": "integer"
              },
              "date": {
                  "type": "string"
              },
              "id": {
                  "type": "string"
              },
              "location": {
                  "type": "string"
              },
              "name": {
                  "type": "string"
              },
              "organization": {
                  "type": "string"
              },
              "partner_id": {
                  "type": "integer"
              },
              "price": {
                  "type": "number"
              },
              "rating": {
                  "type": "string"
              }
          }
      },
      "usecase.ListEventsOutputDTO": {
          "type": "object",
          "properties": {
              "events": {
                  "type": "array",
                  "items": {
                      "$ref": "#/definitions/usecase.EventDTO"
                  }
              }
          }
      },
      "usecase.ListSpotsOutputDTO": {
          "type": "object",
          "properties": {
              "event": {
                  "$ref": "#/definitions/usecase.EventDTO"
              },
              "spots": {
                  "type": "array",
                  "items": {
                      "$ref": "#/definitions/usecase.SpotDTO"
                  }
              }
          }
      },
      "usecase.SpotDTO": {
          "type": "object",
          "properties": {
              "event_id": {
                  "type": "string"
              },
              "id": {
                  "type": "string"
              },
              "name": {
                  "type": "string"
              },
              "reserved": {
                  "type": "boolean"
              },
              "status": {
                  "type": "string"
              },
              "ticket_id": {
                  "type": "string"
              }
          }
      },
      "usecase.TicketDTO": {
          "type": "object",
          "properties": {
              "id": {
                  "type": "string"
              },
              "price": {
                  "type": "number"
              },
              "spot_id": {
                  "type": "string"
              },
              "ticket_type": {
                  "type": "string"
              }
          }
      }
  }
}